+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### v2.0.0
 * FEATURE | Added SFX! Yay!
+* FEATURE | Added playcount counter.
 * FIX | Now the level counter can't roll past end level.
 
 ### v1.7.0